Refactored file to comply with new folder structure
