+++ src/util/launcher/launcher.c
@@ -3,8 +3,8 @@
 #include <stdio.h>
 #include <stdlib.h>
 
-#include "log/proginfo.h"
-#include "log/log.h"
+#include "util/log/proginfo.h"
+#include "util/log/log.h"
 
 void launch(int argc, char *argv[]) {
   // TODO implement
